chore: reorganizado DTOs em pastas separadas para contexto de entrada ou saida
