+++ src/main/java/br/com/fatecmc/tecladostorelesbackend/presentation/dtos/input/EnderecoInputDTO.java
@@ -1,0 +1,4 @@
+package br.com.fatecmc.tecladostorelesbackend.presentation.dtos.input;
+
+public class EnderecoInputDTO {
+}
