+++ src/main/java/br/com/fatecmc/tecladostorelesbackend/presentation/dtos/output/CartaoDTO.java
@@ -1,6 +1,7 @@
-package br.com.fatecmc.tecladostorelesbackend.presentation.dtos;
+package br.com.fatecmc.tecladostorelesbackend.presentation.dtos.output;
 
 import br.com.fatecmc.tecladostorelesbackend.domain.models.enums.cartoes.Bandeiras;
+import br.com.fatecmc.tecladostorelesbackend.presentation.dtos.BaseDTO;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,7 +11,7 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class CartaoDTO extends BaseDTO{
+public class CartaoDTO extends BaseDTO {
     private Boolean preferencial;
     private String numeros;
     private Bandeiras bandeira;
