+++ src/main/java/br/com/fatecmc/tecladostorelesbackend/presentation/dtos/output/EnderecoDTO.java
@@ -1,7 +1,8 @@
-package br.com.fatecmc.tecladostorelesbackend.presentation.dtos;
+package br.com.fatecmc.tecladostorelesbackend.presentation.dtos.output;
 
 import br.com.fatecmc.tecladostorelesbackend.domain.models.Cliente;
 import br.com.fatecmc.tecladostorelesbackend.domain.models.enums.enderecos.TiposEndereco;
+import br.com.fatecmc.tecladostorelesbackend.presentation.dtos.BaseDTO;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -11,7 +12,7 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class EnderecoDTO extends BaseDTO{
+public class EnderecoDTO extends BaseDTO {
     private String nomeEndereco;
     private TiposEndereco tipoEndereco;
     private String tipoResidencia;
